Move test env classes to `test-fixtures`
